+++ cpp/SampleIME/cbindgen/numberkey_windows.h
@@ -5,7 +5,7 @@
 
 extern "C" {
 
-int32_t index_from_number_key(uint32_t vkey);
+int16_t index_from_number_key(uint16_t vkey);
 
 uint32_t number_key_label_at(uint32_t index);
